summarization/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.23.2
pydantic==2.4.2
python-multipart==0.0.20

# RAG system dependencies
openai==1.3.0
tiktoken==0.5.1
sentence-transformers==2.2.2
torch>=2.0.0
backoff==2.2.1
numpy==1.24.3

# Supabase integration
supabase==1.0.3
pyjwt==2.8.0

# Utilities
requests==2.31.0
python-dotenv==1.0.0
docopt==0.6.2
rich==13.4.2
pyperclip==1.8.2